Create Code
The code for this project is written in C++ using Object-Oriented Programming (OOP) 
principles and developed in the Arduino IDE environment. Arduino IDE is an open-source 
platform that supports programming of Arduino-compatible boards using simplified C/C++ 
syntax. Object-oriented concepts such as classes and encapsulation were utilized to structure the 
code in a modular and reusable format, enhancing readability and maintainability.